feat: rename files for consistency and clarity
