+++ frontend/app/[locale]/(routes)/events/_components/EventPageNav.tsx
