+++ frontend/app/[locale]/(routes)/events/hackathons/hackathon2022/_data/partnerImages.ts
