+++ frontend/app/[locale]/(routes)/events/hackathons/hackathon2023/_components/HowDidItGo.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { SkeletonLoader } from '@/components/hackathon/skeletons/skeleton-loader';
+import { SkeletonLoader } from '@/components/hackathon/skeletons/SkeletonLoader';
 import Image from 'next/image';
 import Link from 'next/link';
 import { useState } from 'react';
