+++ frontend/components/layout/FriendsSection.tsx
