+++ frontend/components/layout/HeroHeader.tsx
@@ -1,7 +1,7 @@
 'use client';
 import Image from 'next/image';
 import { useState } from 'react';
-import { SkeletonLoader } from '../hackathon/skeletons/skeleton-loader';
+import { SkeletonLoader } from '../hackathon/skeletons/SkeletonLoader';
 
 interface HeroHeaderProps {
   firstDesc: string;
