+++ frontend/components/layout/InfoSection.tsx
